Customer transaction
Upserting a customer requires multiple prisma queries to associate relations with the customer. Now these are done in one transaction, so that any errors don't cause partial data to be added. Previously, creating a new customer from the admin console could run into this error, if the user you're trying to create is using an existing phone number or email